845 file attachment fix

+++ app/logic/fileHandler.py
@@ -12,41 +12,55 @@ class FileHandler:
         if courseId:
             self.path = os.path.join(self.path, app.config['files']['course_attachment_path'], str(courseId))
         elif eventId:
-            self.path = os.path.join(self.path, app.config['files']['event_attachment_path'], str(eventId))
+            # eventID is not included in the path, because it is now a part of the attachment filename.
+            self.path = os.path.join(self.path, app.config['files']['event_attachment_path'])
+        
         try:
-            os.makedirs(self.path)
+            extraDir = str(self.eventId) if self.eventId else ""
+            os.makedirs(os.path.join(self.path, extraDir))
         except:
             print("Directory exists.")
 
-    def getFileFullPath(self, newfile = None):
+    def getFileFullPath(self, newfilename = ''):
         """
         This creates the directory/path for the object from the "Choose File" input in the create event and edit event.
         :returns: directory path for attachment
         """
+
+        # Added the eventID of the first recurring event to track the file path for subsequent recurring events.
         try:
             # tries to create the full path of the files location and passes if
             # the directories already exist or there is no attachment
-            filePath=(os.path.join(self.path, newfile.filename))
+            filePath=(os.path.join(self.path, newfilename))
         except AttributeError:  # will pass if there is no attachment to save
             pass
         except FileExistsError:
             pass
+
         return filePath
 
-    def saveFiles(self):
+    def saveFiles(self, saveOriginalFile=None):
         """ Saves the attachment in the app/static/files/eventattachments/ or courseattachements/ directory """
         try:
             for file in self.files:
                 if self.eventId:
-                    isFileInEvent = AttachmentUpload.select().where(AttachmentUpload.event == self.eventId, AttachmentUpload.fileName == file.filename).exists()
+                    attachmentName = str(saveOriginalFile.id) + "/" +  file.filename
+
+                    # isFileInEvent checks if the attachment exists in the database under that eventId and filename.
+                    isFileInEvent = AttachmentUpload.select().where(AttachmentUpload.event == self.eventId,
+                                                                    AttachmentUpload.fileName == attachmentName).exists()
                     if not isFileInEvent:
-                        AttachmentUpload.create(event = self.eventId, fileName = file.filename)
-                        file.save(self.getFileFullPath(newfile = file)) # saves attachment in directory
+                        AttachmentUpload.create(event = self.eventId, fileName = attachmentName)
+
+                        # Only save the file if our event is on its own, or the first of a recurring series
+                        if saveOriginalFile and saveOriginalFile.id == self.eventId:       
+                            file.save(self.getFileFullPath(newfilename = attachmentName))
+ 
                 elif self.courseId:
                     isFileInCourse = AttachmentUpload.select().where(AttachmentUpload.course == self.courseId, AttachmentUpload.fileName == file.filename).exists()
                     if not isFileInCourse:
                         AttachmentUpload.create(course = self.courseId, fileName = file.filename)
-                        file.save(self.getFileFullPath(newfile = file)) # saves attachment in directory
+                        file.save(self.getFileFullPath(newfilename = file.filename))
         except AttributeError: # will pass if there is no attachment to save
             pass
 
@@ -61,10 +75,10 @@ class FileHandler:
         """
         Deletes attachmant from the app/static/files/eventattachments/ or courseattachments/ directory
         """
-        try:
-            file = AttachmentUpload.get_by_id(fileId)
+        file = AttachmentUpload.get_by_id(fileId)
+        file.delete_instance()
+
+        # checks if there are other instances with the same filename in the AttachmentUpload table
+        if not AttachmentUpload.select().where(AttachmentUpload.fileName == file.fileName).exists():
             path = os.path.join(self.path, file.fileName)
             os.remove(path)
-            file.delete_instance()
-        except AttributeError: #passes if no attachment is selected.
-            pass
